Revert "rtc: control default alarm wake up capability"
Our RTC peripheral in PM8941 PMIC does not support disabling RTC alarm wakeup.
This property is enabled only on msm-5.4 and msm-5.15 at the moment for iot
and vm platforms which indicates that it is clearly not for our SDM660 mobile
platform. We are never going to enable "disable-alarm-wakeup" at dts. So, drop
this useless patch.

This reverts commit 0ee97ccec85bcfa0b268038439e8dc8a952522e9.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: RyuujiX <saputradenny712@gmail.com>